workloadmeta: extract SBOM utilities into separate package
