+++ comp/core/workloadmeta/collectors/sbomutil/doc.go
@@ -1,0 +1,7 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2023-present Datadog, Inc.
+
+// Package sbomutil contains SBOM utility functions for workload metadata collectors
+package sbomutil
