+++ comp/core/workloadmeta/collectors/sbomutil/image_util_test.go
@@ -5,7 +5,7 @@
 
 //go:build trivy
 
-package util
+package sbomutil
 
 import (
 	"reflect"
